Move options to configuration file
Moved the global configuration options for qubes-keepass to a separate
configuration file. This makes it easier to update the script while
still having a persistent configuration.

The configuration format was chosen to be an INI file. Probably not the
best format, but configparser is within the python standard library and
can be used without further dependencies.