Apply uuid cache before copying credentials
After selecting a credential, the credential was first copied to the
qube and afterwards put into the cache. However, since copying contains
a ten seconds delay in which the clipboard is cleared, it took 10
seconds for the credential to appear within the cache. This prevented
smart_search from working correctly.