Prevent vault from starting if not already strated
When using qubes-keepass while the vault VM has not been started, it
starts automatically. This kind of qube startup is known to cause
problems with the dbus integration of KeePassXC. It is probably because
qubes-keepass is querying the secret-service while KeePassXC is not
running which leads to a startup of gnome-keyring. To prevent this, this
commit prevents the vault VM from being started if it was not already
started.